Write commit files to their corresponding branch directories
Before this change, we were writing commits to the .git directory in
the HTML output directory, and hard-linking commit files from the
branch-specific commit directory to the corresponding ones in the .git
directory.

This doesn't work well, as HTML files for commits which are in
multiple branches in the .git directory will have the preamble
referencing only one of the branches, thus when listed from branches
which are not the one in the preamble, the preamble would be wrong.

Fix this by dropping the hard-linking strategy altogether and writing
HTML commit files directly into the branch directory.  There'll be
duplicates in terms of contents of commits, but the preambles will be
consistent.